### 2017-06-08 #287 @whedon
[PRE REVIEW]: ClimDown: Climate Downscaling in R
**Submitting author:** @jameshiebert (<a href="http://orcid.org/0000-0002-4171-9586">James Hiebert</a>)
**Repository:** <a href="https://github.com/pacificclimate/ClimDown" target ="_blank">https://github.com/pacificclimate/ClimDown</a>
**Version:** 1.0.3
