### 2017-07-21 #287 @kyleniemeyer
@arfon @karthik Is this another candidate for going through review via rOpenSci?
