### 2017-07-21 #287 @karthik
@kyleniemeyer Unfortunately this is not a fit for ropensci. So this should proceed as is with JOSS>
