### 2017-07-21 #287 @kyleniemeyer
@karthik ok, perhaps one of the R-focused @openjournals/joss-editors can take it them 😋
