### 2017-08-11 #360 @whedon
[REVIEW]: ClimDown: Climate Downscaling in R
**Submitting author:** @jameshiebert (<a href="http://orcid.org/0000-0002-4171-9586">James Hiebert</a>)
**Repository:** <a href="https://github.com/pacificclimate/ClimDown" target ="_blank">https://github.com/pacificclimate/ClimDown</a>
**Version:** v1.0.7
