### 2017-09-28 #360 @mdsumner
Here is my review notes for the specific checks.

## Installation

There's no notes in the readme for installing from either CRAN or Github.

Sufficient code to install is

```R
## install.packages("devtools")
devtools::install_github("pacificclimate/ClimDown")
```

For R CMD check include

```R
install.packages("RUnit")
```

A .travis.yml check would be good, since there are system dependencies (NetCDF) and having
a known build script is helpful.

## General checks

* License: GPL-3 in DESCRIPTION (not in LICENSE file)

* Release version 1.0.3 is in a branch (not in recent history of trunk)

## Functionality

* No instructions for installation in the readme, but there is in the CONTRIBUTING.rst and the installation as per Github/R Packages standard was fine. Since NetCDF can be tricky, having
Linux and OSX install instructions would be helpfule.

* I don't find it clear what the purpose of the package is from the README or documentation, I think
this needs more explanation of the motivation.

* Performance I think is hard to demonstrate since the application is so specific, and it's very
deeply wrapped up in parallelization schemes. The README should mention the systems on which this
has been applied and whether it's worth trying to use this on other systems.

## Documentation

The documentation describes the overall process, but not the mechanics of using the package much. There aren't examples beyond an uncommented sequence in `parallelization` and in the exec/ scripts.

There are full examples are in the tests, and these are pretty comprehensive - I think some of the code here should be used as examples in the documentation.  It needs to be made clearer up front that the main interface for this package is to run via the scripts in exec/.

The paper.md mentions '... high-level wrapper functions that perform each of three downscaling steps: CI, CA, and QDM, as well as one wrapper that runs the entire BCCAQ pipeline.'

The actual functions are

```
bccaq.netcdf.wrapper

ca.netcdf.wrapper
ci.netcdf.wrapper
qdm.netcdf.wrapper

rerank.netcdf.wrapper
```

with the first being the 'one wrapper', the middle three are the downscaling steps and the third is a specialized sort to be sufficiently fast. Each of these functions are so-named wrappers because they don't actually perform in a functional sense, but to be invoked from the command line with the scripts
that are in /exec/.  I think this workflow should be made clearer, and they can be easily summarized on the front page/s (README, package-doc) to explain what they do in one place.

`bccaq`, `ci` and `qdm` accept input as two input file names and provide output as a side effect written to a third file in the form `obs.file, gcm.file, out.file, varname='tasmax'`.

`ca` take the same analogous input files but return a list of vectors.

`rerank` is an optimization, does it need to be exported?

* Community input: there is the Github issues in the DESCRIPTION.

## Software paper

The motivation for who the users of this package is useful for is not made clear. Is this to publish
a set of working scripts for the PCIC, or is it also intended for general users of climate data to pick up this package and use?
