### 2017-12-07 #360 @jameshiebert
Yes, thanks @mdsumner for taking the time to review. I'll try to put together a detailed response early next week.
