### 2017-12-12 ed0860d
Started addressing JOSS review
