+++ README.rst
@@ -3,10 +3,11 @@ What is ClimDown?
 
 "ClimDown" is a Climate Downscaling package for the `R statistical
 programming language`. It was written at the `Pacific Climate Impacts
-Consortium`_ (PCIC) with support from `Environment Canada`_.
+Consortium`_ (PCIC) with support from `Environment and Climate Change
+Canada`_.
 
-The package provides routines for downscaling coarse scale global
-climate model (GCM) output to a fine spatial resolution.
+The package provides routines for statistical downscaling of coarse
+scale global climate model (GCM) output to a fine spatial resolution.
 
 PCIC's suite of routines include several different (yet related)
 downscaling techniques:
@@ -18,7 +19,25 @@ downscaling techniques:
 
 See the corresponding scientific literature for more details on the
 pros and cons of each downscaling method.
-  
+
   .. _R statistical programming language: http://www.r-project.org/
   .. _Pacific Climate Impacts Consortium: https://pacificclimate.org/
-  .. _Environment Canada: http://ec.gc.ca/
+  .. _Environment and Climate Change Canada: http://ec.gc.ca/
+
+
+Installation
+------------
+
+You can install the latest `ClimDown release from CRAN`_ using the R
+interpreter: ``> install.packages('climdex.pcic')``
+
+.. _ClimDown release from CRAN: http://cran.r-project.org/web/packages/ClimDown/index.html
+
+If you are interested in a development version or a specific release
+of ClimDown, you can use the `devtools` package as an installation
+alternative.
+
+     > install.packages('devtools')
+     > devtools::install_github("pacificclimate/ClimDown", ref="release")
+     # Or
+     > devtools::install_github("pacificclimate/ClimDown", ref="1.0.1")
