### 2017-12-14 #360 @jameshiebert
Thanks for the detailed review and excellent comments. I have clarified a few things inline below, and have noted where we need to make changes to address your concerns.

> There's no notes in the readme for installing from either CRAN or Github.

Fixed [in branch](https://github.com/pacificclimate/ClimDown/tree/joss-review).

> A .travis.yml check would be good, since there are system dependencies (NetCDF) and having a known build script is helpful.

You must have missed it, because we have been running Travis builds for a bit over a year. You can see the config [here](https://github.com/pacificclimate/ClimDown/blob/master/.travis.yml) and the [build output here](https://travis-ci.org/pacificclimate/ClimDown/builds).

> License: GPL-3 in DESCRIPTION (not in LICENSE file)

The GPL-3 license was in a `COPYING` file which is standard in Open Source projects. To address your comment, we have now since included `LICENSE` as a symbolic link to `COPYING`.

> Release version 1.0.3 is in a branch (not in recent history of trunk)

All 16 releases are tagged in the source control, for example:

https://github.com/pacificclimate/ClimDown/releases/tag/1.0.3

Our practice is to maintain a ["release" branch](https://github.com/pacificclimate/ClimDown/tree/release), that accepts merges from master, and maintains any automatically generated files required for installation (e.g. all the .Rd doc files).

> Since NetCDF can be tricky, having Linux and OSX install instructions would be helpful.

Good suggestion. We have added a description of system deps and linked to install instructions [in branch](https://github.com/pacificclimate/ClimDown/tree/joss-review).

> I don't find it clear what the purpose of the package is from the
> README or documentation, I think this needs more explanation of the
> motivation.

Good suggestion. We'll fix this.

> Performance I think is hard to demonstrate since the application is
> so specific, and it's very deeply wrapped up in parallelization
> schemes. The README should mention the systems on which this has
>  been applied and whether it's worth trying to use this on other
>  systems.

Good suggestion. We'll fix this.

> There are full examples are in the tests, and these are pretty
> comprehensive - I think some of the code here should be used as
> examples in the documentation. It needs to be made clearer up front
> that the main interface for this package is to run via the scripts
> in exec/.

Good suggestion. We'll fix this.

> third is a specialized sort to be sufficiently fast.
> ...
> rerank is an optimization, does it need to be exported?

No, `rerank` is not simply an optimization; it is imperative to the credibility of the downscaling process. We will explain that in better detail.

> The motivation for who the users of this package is useful for is
> not made clear. Is this to publish a set of working scripts for the
> PCIC, or is it also intended for general users of climate data to
> pick up this package and use?

General use. We'll try to motivate it better.
