### 2017-12-14 3bc66c3
Added README section about performance
