### 2017-12-14 43636bc
Clarifies downscaling steps in docs
