### 2017-12-14 a22e880
Describes what/why of downscaling in the README
