### 2017-12-15 3889642
Adds examples to package docs
