### 2018-01-12 #360 @mdsumner
Excellent, thanks for all this.

Just a few notes.

The devtools installation can use this syntax now, which you might prefer:

```{r}
devtools::install_github("pacificclimate/ClimDown@release")
```

I did miss the .travis.yml, and now realize why - I'm used to seeing "badges" in github readmes, which you might consider adding - but I'll remember not to assume travis will always include these ...

I think you've addressed all my comments very well. Thanks!

I've confirmed all the check boxes in the list above.
