### 2018-02-01 5d11a9a
Release of v1.0.4
