+++ man/ClimDown.Rd
@@ -4,6 +4,7 @@
 \name{ClimDown}
 \alias{ClimDown}
 \alias{ClimDown-package}
+\alias{ClimDown-package}
 \title{ClimDown: PCIC's daily Climate Downscaling library}
 \description{
 This package includes PCIC's routines and techniques for
@@ -12,11 +13,12 @@ spatial resolution.
 }
 \details{
 At present, the package only exports high-level wrapper function
-that perform each of three downscaling steps: CI, CA, and QDM, as
-well as one wrapper that runs the entire pipeline: BCCAQ. In
-general, each wrapper simply takes four arguments: GCM file,
-gridded observation file, output file, and variable name. However,
-see the specific function documentation for specifics.
+that perform each of three downscaling steps: CI, CA, QDM, and
+rerank as well as one wrapper that runs the entire pipeline:
+BCCAQ. In general, each wrapper simply takes four arguments: GCM
+file, gridded observation file, output file, and variable
+name. However, see the specific function documentation for
+specifics.
 
 The package also provides five wrapper scripts that allow the user
 to run each step from the command line (plus the whole pipeline)
@@ -27,4 +29,3 @@ Werner, A. T., & Cannon, A. J. (2016). Hydrologic extremes  - an intercomparison
 }
 \keyword{climate}
 \keyword{downscaling}
-
