+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: ClimDown
-Version: 1.0.3
-Date: 2017-06-07
+Version: 1.0.4
+Date: 2018-02-01
 Title: Climate Downscaling Library for Daily Climate Model Output
 Authors@R: c(person("Alex Cannon", role = "aut", email = "acannon@uvic.ca"),
 	     person("James Hiebert", role = c("aut", "cre"), email = "hiebert@uvic.ca"),
@@ -29,4 +29,4 @@ License: GPL-3 | file LICENSE
 URL: https://www.r-project.org
 LazyData: yes
 BugReports: https://github.com/pacificclimate/ClimDown
-RoxygenNote: 5.0.1
+RoxygenNote: 6.0.1
