### 2018-02-01 8dff536
Fixes typo in paper.md
