### 2018-02-01 #360 @jameshiebert
Ah shoot; sorry about that. Pushed a patch to master/release branches... do I need to generate a new archive?
