### 2018-02-01 #360 @karthik
Yes please update the archive and I'll fix it here, then recompile the PDF.
