### 2018-02-01 b34965b
Release of v1.0.5
