+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: ClimDown
-Version: 1.0.4
+Version: 1.0.5
 Date: 2018-02-01
 Title: Climate Downscaling Library for Daily Climate Model Output
 Authors@R: c(person("Alex Cannon", role = "aut", email = "acannon@uvic.ca"),
