### 2018-02-01 #360 @jameshiebert
Done. 10.5281/zenodo.1164299
