### 2018-02-01 #360 @karthik
Sorry, one more issue. Can you move your references to `paper.bib` and remove them from the paper?
