### 2018-02-01 5d5dfb8
Moves JOSS references into BibTeX
