+++ paper.bib
@@ -1,0 +1,56 @@
+% This file was created with JabRef 2.10.
+% Encoding: ISO8859_1
+
+
+@Article{cannon15,
+  Title                    = {Bias Correction of GCM Precipitation by Quantile Mapping: How Well Do Methods Preserve Changes in Quantiles and Extremes?},
+  Author                   = {Cannon, A. J. and Sobie, S. R. and Murdock, T. Q.},
+  Year                     = {2015},
+  Doi                      = {10.1175/JCLI-D-14-00754.1},
+  Pages                    = {6938-6959},
+  Volume                   = {28(17)},
+
+  Journal                  = {Journal of Climate}
+}
+
+@Article{hunter05,
+  Title                    = {Climatologically aided mapping of daily precipitation and temperature},
+  Author                   = {Hunter, R. D. and Meentemeyer, R. K.},
+  Year                     = {2005},
+  Pages                    = {1501-1510},
+  Volume                   = {44(10)},
+
+  Journal                  = {Journal of Applied Meteorology}
+}
+
+@Article{maurer10,
+  Title                    = {The utility of daily large-scale climate data in the assessment of climate change impacts on daily streamflow in California},
+  Author                   = {Maurer, E. P. and Hidalgo, H. G. and Das, T. and Dettinger, M. D. and Cayan, D. R},
+  Year                     = {2010},
+  Pages                    = {1125-1138},
+  Volume                   = {14(6)},
+
+  Journal                  = {Hydrology and Earth System Sciences}
+}
+
+@Article{schefzik13,
+  Title                    = {Uncertainty quantification in complex simulation models using ensemble copula coupling},
+  Author                   = {Schefzik, R. and Thorarinsdottir, T. L. and Gneiting, T.},
+  Year                     = {2013},
+  Pages                    = {616-640},
+  Volume                   = {28(4)},
+
+  Journal                  = {Statistical Science}
+}
+
+@Article{werner16,
+  Title                    = {Hydrologic extremes - an intercomparison of multiple gridded statistical downscaling methods},
+  Author                   = {Werner, A. T. and Cannon, A. J.},
+  Year                     = {2016},
+  Doi                      = {10.5194/hess-20-1483-2016},
+  Pages                    = {1483-1508},
+  Volume                   = {20(4)},
+
+  Journal                  = {Hydrology and Earth System Sciences}
+}
+
