### 2018-02-01 b1d6f14
Release of v1.0.6
