### 2018-02-02 #360 @karthik
@arfon Could you check to make sure the bib parses correctly in the PDF?
