### 2018-02-14 #360 @arfon
@jameshiebert - could you make sure you're citing the references in the body of your paper please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
