### 2018-02-14 #360 @jameshiebert
Sure. Is there way that I can test it out locally before committing (e.g. with pandoc or something)? What's the command that you use for the article proof?
