### 2018-02-14 809bade
Ensure references are cited in JOSS paper
