+++ paper.md
@@ -34,14 +34,15 @@ The ClimDown R package publishes the routines and techniques of the
 fine scale spatial resolution.
 
 PCIC's overall downscaling algorithm is named Bias-corrected
-constructed analogues with quantile mapping (BCCAQ). BCCAQ is a hybrid
-downscaling method that combines outputs from Climate Analogues (CA)
-and quantile mapping at the fine-scale resolution.  First, the CA and
-Climate Imprint (CI) plus quantile delta mapping (QDM) algorithms are
-run independently. BCCAQ then combines outputs from the two by taking
-the daily QDM outputs at each fine-scale grid point and reordering
-them within a given month according to the daily CA ranks, i.e., using
-a form of Empirical Copula Coupling.
+constructed analogues with quantile mapping (BCCAQ)
+[@cannon15; @werner16]. BCCAQ is a hybrid downscaling method that
+combines outputs from Climate Analogues (CA) [@maurer10] and quantile
+mapping at the fine-scale resolution.  First, the CA and Climate
+Imprint (CI) [@hunter05] plus quantile delta mapping (QDM) [@cannon15]
+algorithms are run independently. BCCAQ then combines outputs from the
+two by taking the daily QDM outputs at each fine-scale grid point and
+reordering them within a given month according to the daily CA ranks,
+i.e., using a form of Empirical Copula Coupling [@schefzik13].
 
 The package exports high-level wrapper functions that perform each of
 three downscaling steps: CI, CA, and QDM, as well as one wrapper that
