### 2018-02-23 b4a5c85
Makes several minor corrections from Alex Cannon
