+++ paper.md
@@ -9,7 +9,7 @@ authors:
   - name: James Hiebert
     orcid: 0000-0002-4171-9586
     affiliation: 1
-  - name: Alex Cannon
+  - name: Alex J. Cannon
     affiliation: 2
   - name: Trevor Murdock
     affiliation: 1
@@ -20,7 +20,7 @@ authors:
 affiliations:
   - name: Pacific Climate Impacts Consortium
     index: 1
-  - name: Environment and Climate Change Canada
+  - name: Climate Research Division, Environment and Climate Change Canada
     index: 2
 date: 1 February 2018
 bibliography: paper.bib
@@ -34,9 +34,9 @@ The ClimDown R package publishes the routines and techniques of the
 fine scale spatial resolution.
 
 PCIC's overall downscaling algorithm is named Bias-corrected
-constructed analogues with quantile mapping (BCCAQ)
+constructed analogues with quantile mapping reordering (BCCAQ)
 [@cannon15; @werner16]. BCCAQ is a hybrid downscaling method that
-combines outputs from Climate Analogues (CA) [@maurer10] and quantile
+combines outputs from Constructed Analogues (CA) [@maurer10] and quantile
 mapping at the fine-scale resolution.  First, the CA and Climate
 Imprint (CI) [@hunter05] plus quantile delta mapping (QDM) [@cannon15]
 algorithms are run independently. BCCAQ then combines outputs from the
