### 2018-02-23 #360 @jameshiebert
@arfon References are cited in the paper and now included in the bibliography. Do you need me to great a new DOI/archive or can you work off my master branch?
