### 2018-02-24 #360 @arfon
> @arfon References are cited in the paper and now included in the bibliography. Do you need me to create a new DOI/archive or can you work off my master branch?

Ideally we'd like there to be a new release that matches what was the finalized version of the repository when accepted.
