### 2018-02-26 #360 @jameshiebert
@afron OK 10.5281/zenodo.1184647 is the new archive for v1.0.7.
